+++ bin/monorepo-npm-check-updates/index.mjs
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 'use strict'
-import { monorepo_npm_check_updates_cli } from '../src/index.js'
+import { monorepo_npm_check_updates_cli } from '../../src/index.js'
 monorepo_npm_check_updates_cli()
 	.then(()=>process.exit(0))
 	.catch(()=>process.exit(1))