+++ bin/update-npm-packages.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 DIR="$(dirname $0)"
 TEMP=$(mktemp)
-"$DIR/npm-check-updates--monorepo.mjs" > $TEMP
+"$DIR/monorepo-npm-check-updates/index.mjs" -o $TEMP
 echo "$TEMP"
 cat "$TEMP"
 cat "$TEMP" | "$DIR/package-manifest-changeset.sh"